Allow for complex relations in Query
Currently it is only possible to do `field => value` map queries that result in an AND condition of all fields. This PR introduces a backwards compatible way to do more complex relations and value comparisons. The query `[ $field => $value ]` is effectively now shorthand for:

```
Query( [
	'relation' => 'AND',
	'fields' => [ $field => $value ],
] );
```

Where `fields` can either be a map of fields, or multidimensional array of clauses following the same `[ relation => string, fields => array ]` recursive structure.

Similary, the same applies for fiel values, where `$value` can either be a scalar that is essentially shorthand for `[ compare => '=', value => $value ]`. As you can see, supported comparisons are currently `>, >=, <, <=, =, ~=, LIKE`. I'd suggest we add `IN` / `NOT IN` to this, and do away with the (not implemented) `$field__in` style keys.